refactor: drop V2 suffixes and collapse legacy spec/error modules
The lockfile-v7 refactor introduced parallel "V2" types alongside the
old ones. Now that the transition is over, consolidate:

* `SourceMetadataSpecV2` -> `SourceMetadataSpec` (the Key spec wins the
  canonical name); legacy reporter-shaped `SourceMetadataSpec` is
  renamed to `SourceMetadataReporterSpec` and moved to `reporter.rs`.
* `SourceBuildSpecV2` -> `SourceBuildSpec`.
* Legacy `SourceRecordSpec` (reporter-shaped) -> `SourceRecordReporterSpec`,
  moved to `reporter.rs`. Unused `ResolvedSourceRecord` deleted.
* `SolveCondaSpec` (Key) and `SolveCondaEnvironmentSpec` (full solver
  input) kept as distinct semantic names: they live at different layers
  (the Key fetches binary repodata before constructing the env spec).

Module collapses: the contents of `source_build/`, `source_metadata/`,
`source_record/`, and `solve_pixi/` were just errors plus a cycle
helper, so:

* All errors (`SourceBuildError`, `SourceMetadataError`,
  `SourceRecordError`, `SolvePixiEnvironmentError`,
  `MissingChannelError`) now live in a top-level `errors.rs`.
* `Cycle` / `CycleEnvironment` move to top-level `cycle.rs`; snapshot
  file renamed to match the new module path.
* The `SourceMetadata` records struct moves to `keys/source_metadata.rs`
  next to `SourceMetadataKey`.
* `check_missing_channels` helper moves to
  `keys/solve_pixi_environment.rs` (its only caller).

Cache extraction to a separate `pixi_build_cache` crate was investigated
and skipped: `cache/build_backend_metadata.rs` and `cache_dirs.rs`
depend on `BuildEnvironment`, `CanonicalSourceCodeLocation`,
`ConfigurationHash`, and `ProjectModelHash` from the dispatcher.
Extracting cleanly would require either cascading those types into the
new crate or having the cache crate depend on the dispatcher (circular
dep risk). Not worth doing now.
