+++ crates/pixi_cli/src/publish.rs
@@ -12,7 +12,7 @@ use pixi_build_frontend::BackendOverride;
 use pixi_command_dispatcher::{
     BuildBackendMetadataSpec, BuildEnvironment, BuildProfile, CacheDirs, ComputeResultExt,
     EnvironmentRef, EnvironmentSpec, EphemeralEnv,
-    keys::{ResolveSourcePackageKey, ResolveSourcePackageSpec, SourceBuildKey, SourceBuildSpecV2},
+    keys::{ResolveSourcePackageKey, ResolveSourcePackageSpec, SourceBuildKey, SourceBuildSpec},
 };
 use pixi_config::{Config, ConfigCli};
 use pixi_core::{WorkspaceLocator, environment::sanity_check_workspace};
@@ -279,7 +279,7 @@ pub async fn execute(args: Args) -> miette::Result<()> {
 
     for record in resolved_records {
         let record = Arc::unwrap_or_clone(record);
-        let build_spec = SourceBuildSpecV2 {
+        let build_spec = SourceBuildSpec {
             record: Arc::new(record.into()),
             channels: channels.clone(),
             exclude_newer: None,
