+++ crates/pixi_command_dispatcher/src/keys/source_metadata.rs
@@ -13,16 +13,25 @@ use std::{
 use derive_more::Display;
 use pixi_build_types::procedures::conda_outputs::CondaOutput;
 use pixi_compute_engine::{ComputeCtx, Key};
-use pixi_record::PinnedSourceSpec;
+use pixi_record::{PinnedSourceSpec, SourceRecord};
 use pixi_spec::SourceLocationSpec;
 use rattler_conda_types::PackageName;
 use tracing::instrument;
 
 use crate::{
     BuildBackendMetadataKey, BuildBackendMetadataSpec, EnvironmentRef, PackageNotProvidedError,
-    build::PinnedSourceCodeLocation, source_checkout::SourceCheckoutExt,
-    source_metadata::SourceMetadataError,
+    SourceMetadataError, build::PinnedSourceCodeLocation, source_checkout::SourceCheckoutExt,
 };
+
+/// The result of resolving source metadata for all variants of a package.
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
+pub struct SourceMetadata {
+    /// Manifest and optional build source location for this metadata.
+    pub source: PinnedSourceCodeLocation,
+
+    /// The metadata that was acquired from the build backend.
+    pub records: Vec<Arc<SourceRecord>>,
+}
 
 /// Input to [`SourceMetadataKey`].
 ///
@@ -30,7 +39,7 @@ use crate::{
 /// `ctx.pin_and_checkout` as its first step. Each `SourceMetadataKey`
 /// runs in its own spawned task, so concurrent fan-out is safe.
 #[derive(Debug, Clone, Eq, PartialEq, Hash)]
-pub struct SourceMetadataSpecV2 {
+pub struct SourceMetadataSpec {
     /// The package whose outputs we want.
     pub package: PackageName,
     /// Unpinned source location; pinned inside the compute unless
@@ -67,10 +76,10 @@ pub struct SourceOutputs {
 /// Dedups on `(package, source_location, preferred_build_source, env_ref)`.
 #[derive(Clone, Debug, Display)]
 #[display("{}/{}", _0.package.as_source(), _0.source_location)]
-pub struct SourceMetadataKey(pub Arc<SourceMetadataSpecV2>);
+pub struct SourceMetadataKey(pub Arc<SourceMetadataSpec>);
 
 impl SourceMetadataKey {
-    pub fn new(spec: SourceMetadataSpecV2) -> Self {
+    pub fn new(spec: SourceMetadataSpec) -> Self {
         Self(Arc::new(spec))
     }
 }
